add UserFeed component
